lurk, fcomm, clutch build
