+++ clutch/Cargo.toml
@@ -12,6 +12,7 @@ repository = "https://github.com/lurk-lab/lurk-rs"
 anyhow = { workspace = true }
 blstrs =  { workspace = true }
 clap = "4.1.8"
+fcomm = { path = "../fcomm" }
 ff = "0.13"
 lurk = { path = "../" }
 pasta_curves = { workspace = true, features = ["repr-c", "serde"] }
