+++ fcomm/Cargo.toml
@@ -12,8 +12,10 @@ name = "fcomm"
 path = "src/bin/fcomm.rs"
 
 [dependencies]
+anyhow = { workspace = true }
 base64 = { workspace = true }
 bellperson = { workspace = true }
+bincode = { workspace = true }
 blstrs = { workspace = true }
 clap = { version = "3.2", features = ["derive"] }
 clap-verbosity-flag = "1.0"
@@ -25,6 +27,8 @@ once_cell = { workspace = true }
 pairing = { workspace = true }
 pasta_curves = { workspace = true, features = ["repr-c", "serde"] }
 pretty_env_logger = { workspace = true }
+proptest = "1.1.0"
+proptest-derive = "0.3.0"
 rand = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
